Quick cleanup/organization of ukspec
